+++ adot/collector/lambdacomponents/default.go
@@ -15,10 +15,13 @@
 
 package lambdacomponents
 
+//github.com/aws-observability/aws-otel-lambda/adot/collector/lambdacomponents
+
 import (
+	"log"
+
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/awsemfexporter"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/awsxrayexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusexporter"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusremotewriteexporter"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/sigv4authextension"
 	"go.opentelemetry.io/collector/exporter"
@@ -26,6 +29,7 @@ import (
 	"go.opentelemetry.io/collector/exporter/otlpexporter"
 	"go.opentelemetry.io/collector/exporter/otlphttpexporter"
 	"go.opentelemetry.io/collector/extension"
+	"go.opentelemetry.io/collector/featuregate"
 	"go.opentelemetry.io/collector/otelcol"
 	"go.opentelemetry.io/collector/receiver"
 	"go.opentelemetry.io/collector/receiver/otlpreceiver"
@@ -57,7 +61,6 @@ func Components() (
 	exporters, err := exporter.MakeFactoryMap(
 		awsxrayexporter.NewFactory(),
 		awsemfexporter.NewFactory(),
-		prometheusexporter.NewFactory(),
 		prometheusremotewriteexporter.NewFactory(),
 		loggingexporter.NewFactory(),
 		otlpexporter.NewFactory(),
@@ -72,6 +75,14 @@ func Components() (
 		Receivers:  receivers,
 		Exporters:  exporters,
 	}
+
+	// TODO: remove after ADOT Collector v0.30.0 is released
+	if err := featuregate.GlobalRegistry().Set("pkg.translator.prometheus.NormalizeName", false); err != nil {
+		return otelcol.Factories{}, err
+	}
+	log.Printf("attn: users of the prometheusremotewrite exporter please refer to " +
+		"https://github.com/aws-observability/aws-otel-collector/issues/2043 in regards to an ADOT Collector v0.31.0 " +
+		"breaking change")
 
 	return factories, errs
 }