+++ python/build.sh
@@ -11,7 +11,9 @@ popd || exit
 cd ../opentelemetry-lambda/python/src || exit
 ./build.sh
 
-pushd ../sample-apps || exit
+# Build sample app
+
+pushd ../../../python/sample-apps || exit
 ./build.sh
 popd || exit
 
@@ -22,9 +24,9 @@ cd ./build || exit
 pip install opentelemetry-sdk-extension-aws -t python/
 
 # Combine the layers
-unzip -qo layer.zip
-rm layer.zip
-unzip -qo ../../../collector/build/collector-extension.zip
+unzip -qo opentelemetry-python-layer.zip
+rm opentelemetry-python-layer.zip
+unzip -qo ../../../collector/build/opentelemetry-collector-layer-$1.zip
 
 # Use our AWS scripts instead which extend and call OTel Lambda scripts
 mv otel-instrument otel-instrument-upstream-lambda